backend/pyproject.toml
[project]
name = "agent-api"
version = "0.1.0"
description = "Remittance optimization agent service (LangGraph + LangSmith)"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "httpx>=0.27",
  "tenacity>=8.2",
  "slowapi>=0.1",
  "python-jose>=3.3",
  "python-multipart>=0.0.9",
  "langgraph>=0.2",
  "langchain-core>=0.3"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.6"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

